Create an own working copy of read-only sessions
Create an own copy when session is opened and ask for deletion when when
leaving the session view. It would be nice to delete the session
automatically when there are no changes, but it's not easy to know if it
has changed after the page has reloaded and accidental deletion would be
bad. There are remnants of earlier attempts in the code, review before
merging to master.