Make sure there is only one tooltip div
No idea why the style for div.dataset-tooltip seems to be applied only if there is only one element, since Angular 17. This become visible when jumping between Workflow and List. Without styles the divs will stack up over the tab titles and swallow the pointer events (which are also disabled in the styles).